Move strip ansi tests to new file
